+ CHANGES architecture. Adds Load level and restart game support. --> There are some issues using SetCameraAction when loading a new scene. Sometimes Unity will not execute camera handler in the expected way. Restarting Unity usually will fix it, but the behaviour is sometimes unexpected.
