+++ Assets/Prefaps/Starter.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 97c2709f4e0020b4d8426081f119759c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
