+++ Assets/Scenes/0_Startup.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 954632e07305997478bed021ef4e6994
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
