+++ Assets/Scenes/2_Level.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fa48a0687f6199a42abcfb10bd279cc1
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
