+++ Assets/Scripts/GameControl.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7a217259cd93961488382ba39c075d92
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
