+++ Assets/Scripts/GameControl/Script_GameController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 895609a5515b9414da4741068a9fb33c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 100
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
