+++ Assets/Scripts/GameControl/Script_PauseController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ab24ec8b0430f4f49bf88741639dcb7c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
