+++ Assets/Scripts/GameControl/Script_Starter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4ddd7fa6b9b2c374382803476650b73c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 50
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
