+++ Assets/Scripts/GameControl/Script_UIController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 24a10bf4e67510c46893c3f8ba096d9b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
