--- Assets/Scripts/Script_ActivateGame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cb9b5a1ad43c99f4c84bee36f0cebb3d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 100
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
