--- Assets/Scripts/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7a217259cd93961488382ba39c075d92
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
