--- Assets/Scripts/UI/Script_MenuController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7e00f1e729fc06341a8f870848f4a571
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 300
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
