--- Assets/Scripts/UI/Script_PauseController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ab24ec8b0430f4f49bf88741639dcb7c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 200
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
