--- Assets/Scripts/UI/Testing.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5b2eb087849d69e40b20c29785dae76b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
