--- Assets/Scripts/UI/Testing/Script_BoardMessageTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 85d0258ae550eb64fb303e4c82984a6f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 500
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
