+++ Assets/Sound/78528__lg__lecture-12a-small-size.mp3.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9aa8ed44e9e640f4ca1f591efc5f1361
+guid: ebebdce244d2d4b4d9817ba95515b3e3
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
