+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,15 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
+    path: Assets/Scenes/0_Startup.unity
+    guid: 954632e07305997478bed021ef4e6994
+  - enabled: 1
     path: Assets/Scenes/1_Level.unity
-    guid: 560b719e80ca89f4c84b3d822cd26cf2
+    guid: a77d36739b9fc1b478d123b424c18fb7
+  - enabled: 1
+    path: Assets/Scenes/2_Level.unity
+    guid: fa48a0687f6199a42abcfb10bd279cc1
+  - enabled: 1
+    path: Assets/Scenes/3_Level.unity
+    guid: 410fb7a761e6acd4e830f286623b46e5
   m_configObjects: {}
